+++ plone/app/openid/setuphandlers.py
@@ -3,11 +3,20 @@ from plone.app.openid.portlets.login import Assignment as LoginAssignment
 from plone.openid.plugins.oid import addOpenIdPlugin
 from plone.portlets.interfaces import IPortletAssignmentMapping
 from plone.portlets.interfaces import IPortletManager
+from plone.portlets.interfaces import IPortletType
 from Products.CMFCore.utils import getToolByName
 from Products.PlonePAS.browser.info import PASInfoView
 from StringIO import StringIO
 from zope.component import getMultiAdapter
+from zope.component import getSiteManager
 from zope.component import queryUtility
+from zope.component.interfaces import IComponentRegistry
+
+import logging
+
+
+logger = logging.getLogger('plone.app.openid')
+PROFILE_ID = 'plone.app.openid:default'
 
 
 def hasOpenIdPlugin(portal):
@@ -61,3 +70,23 @@ def importVarious(context):
         activatePlugin(site, out, 'openid')
 
     addLoginPortlet(site, out)
+
+
+def registerPortletAddview(context):
+    portal = getToolByName(context, 'portal_url').getPortalObject()
+    sm = getSiteManager(portal)
+    if sm is None or not IComponentRegistry.providedBy(sm):
+        # plone.app.portlets.exportimport.portlets has this same check.
+        logger.info('Cannot register portlet addview - no site manager found.')
+        return
+    if queryUtility(IPortletType, name=''):
+        # For a few years we have registered our portlet with an empty string
+        # as add view.  This is wrong.
+        sm.unregisterUtility(provided=IPortletType, name='')
+        logger.info('Unregistered portlet type with empty addview.')
+    addview = 'portlets.OpenIDLogin'
+    if queryUtility(IPortletType, name=addview) is None:
+        # Apply our default portlets.xml.
+        context.runImportStepFromProfile(PROFILE_ID, 'portlets')
+        logger.info('Applied our portlets.xml to register portlet '
+                    'with correct addview.')